blob/requirements.txt
flask
python-magic
nanoid
Pillow